refactor: useMain hook 분리
